chore: move connect_with... methods to connector
